move product_dict to utils
